Blank Node Support

+++ lib/helperFunctions/addObjectToDataset.ts
@@ -1,21 +1,52 @@
-import { Dataset, Literal, NamedNode, Quad } from "@rdfjs/types";
-import { namedNode, literal, quad } from "@rdfjs/dataset";
+import { BlankNode, Dataset, Literal, NamedNode } from "@rdfjs/types";
+import { namedNode, literal, quad, blankNode } from "@rdfjs/dataset";
 import { ContextUtil } from "../ContextUtil";
-import { ObjectWithId } from "../createSubjectHandler";
-import { deleteValueFromDataset } from "./deleteFromDataset";
+import { getUnderlyingNode } from "../createSubjectHandler";
+
+export type AddObjectItem = {
+  "@id"?: string | NamedNode | BlankNode;
+  [getUnderlyingNode]?: NamedNode | BlankNode;
+} & {
+  [key: string]: AddObjectValue | AddObjectValue[];
+};
+
+export type AddObjectValue = string | boolean | number | AddObjectItem;
+
+function getIdNode(item: AddObjectItem): NamedNode | BlankNode {
+  if (item[getUnderlyingNode]) {
+    return item[getUnderlyingNode] as NamedNode | BlankNode;
+  } else if (!item["@id"]) {
+    return blankNode();
+  } else if (typeof item["@id"] === "string") {
+    return namedNode(item["@id"]);
+  } else {
+    return item["@id"];
+  }
+}
+
+function nodeToSetKey(node: NamedNode | BlankNode): string {
+  if (node.termType === "NamedNode") {
+    return `NamedNode${node.value}`;
+  } else {
+    return `BlankNode${node.value}`;
+  }
+}
 
 export function addObjectValueToDataset(
   dataset: Dataset,
   contextUtil: ContextUtil,
   key: string,
-  visitedObjects: Set<object>,
-  subject: NamedNode,
+  visitedObjects: Set<string>,
+  subject: NamedNode | BlankNode,
   predicate: NamedNode,
-  value: unknown,
+  value: AddObjectValue,
   shouldDeleteOldTriples: boolean
 ): void {
-  let object: NamedNode | Literal;
-  if (
+  // Get the Object Node
+  let object: NamedNode | Literal | BlankNode;
+  if (value == undefined) {
+    dataset.deleteMatches(subject, predicate);
+  } else if (
     typeof value === "string" ||
     typeof value === "boolean" ||
     typeof value === "number"
@@ -27,56 +58,47 @@ export function addObjectValueToDataset(
       object = literal(value.toString(), datatype);
     }
     dataset.add(quad(subject, predicate, object));
-  } else if (
-    value &&
-    typeof value === "object" &&
-    (value as ObjectWithId)["@id"]
-  ) {
-    object = namedNode((value as ObjectWithId)["@id"]);
-    dataset.add(quad(subject, predicate, object));
-    addObjectToDataset(
-      value as ObjectWithId,
-      dataset,
-      contextUtil,
-      visitedObjects,
-      shouldDeleteOldTriples
-    );
   } else {
-    throw Error("Set object does not have an @id");
+    object = getIdNode(value);
+
+    // Delete any triples if the id is the same
+    if (
+      !visitedObjects.has(nodeToSetKey(object)) &&
+      !value[getUnderlyingNode]
+    ) {
+      dataset.deleteMatches(object, undefined, undefined);
+    }
+    dataset.add(quad(subject, predicate, object));
+    if (!value[getUnderlyingNode]) {
+      addObjectToDataset(
+        { ...value, "@id": object } as AddObjectItem,
+        dataset,
+        contextUtil,
+        visitedObjects,
+        shouldDeleteOldTriples
+      );
+    }
   }
 }
 
 export function addObjectToDataset(
-  item: ObjectWithId,
+  item: AddObjectItem,
   dataset: Dataset,
   contextUtil: ContextUtil,
-  visitedObjects: Set<object>,
+  visitedObjects: Set<string>,
   shouldDeleteOldTriples: boolean
 ): void {
-  if (visitedObjects.has(item)) {
+  const subject = getIdNode(item);
+  if (visitedObjects.has(nodeToSetKey(subject))) {
     return;
   }
-  visitedObjects.add(item);
-  const subject = namedNode(item["@id"]);
+  visitedObjects.add(nodeToSetKey(subject));
   Object.entries(item).forEach(([key, value]) => {
     if (key === "@id") {
       return;
     }
     const predicate = namedNode(contextUtil.keyToIri(key));
-    const curQuad: Quad | undefined = dataset
-      .match(subject, predicate)
-      .toArray()[0];
-    // Delete the entire objec if it has the same id
-    if (
-      curQuad &&
-      curQuad.object.termType === "NamedNode" &&
-      typeof value === "object" &&
-      (value as ObjectWithId)["@id"] &&
-      curQuad.object.value === (value as ObjectWithId)["@id"]
-    ) {
-      deleteValueFromDataset(item, key, dataset, contextUtil);
-    }
-    if (shouldDeleteOldTriples) {
+    if (shouldDeleteOldTriples && !item[getUnderlyingNode]) {
       dataset.deleteMatches(subject, predicate);
     }
     if (Array.isArray(value)) {
@@ -100,7 +122,7 @@ export function addObjectToDataset(
         visitedObjects,
         subject,
         predicate,
-        value,
+        value as AddObjectValue,
         true
       );
     }
